move AsciiData* into own files
